centos-7/8, debian 10: images build and deploy
fixes for roles and improvements:
- packer now has correct scopes to build google_cloud_ops_agents
- packer has new image options
- lustre is now installed via archive packages instead of repository
- lmod now can install on debian 10
- cuda task names better reflect OS specific tasks
- cloudagents no longe symlinks fluentd.conf
- added google_cloud_ops_agents role
- mariadb services are stopped and disabled
- common cleaned up to use variables
- cgroups cleaned up to use variables
- python cleaned up to use variables

+++ scripts/setup.py
@@ -659,12 +659,16 @@ def setup_controller():
         cnfdir = Path('/etc/my.cnf.d')
         if not cnfdir.exists():
             cnfdir = Path('/etc/mysql/conf.d')
-        (cnfdir/'mysql_slurm.cnf').write_text("""
+        if not (cnfdir/'mysql_slurm.cnf').exists():
+            (cnfdir/'mysql_slurm.cnf').write_text("""
 [mysqld]
-bind-address = 127.0.0.1
+bind-address=127.0.0.1
+innodb_buffer_pool_size=1024M
+innodb_log_file_size=64M
+innodb_lock_wait_timeout=900
 """)
         util.run('systemctl enable mariadb')
-        util.run('systemctl start mariadb')
+        util.run('systemctl restart mariadb')
 
         mysql = "mysql -u root -e"
         util.run(
